requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
groq==0.4.1
